Rename FontIterator to StyledTextIterator
