+++ embedded-graphics/src/prelude.rs
@@ -2,7 +2,7 @@
 
 pub use super::drawable::Drawable;
 pub use super::drawable::Pixel;
-//pub use super::fonts::Font;
+pub use super::fonts::Font;
 pub use super::geometry::{Dimensions, Point, Size};
 pub use super::image::ImageFile;
 pub use super::pixelcolor::{raw::RawData, GrayColor, PixelColor, RgbColor};
